requirements.txt
pyspark==3.5.8
python-dotenv==1.2.2
psycopg2-binary